+++ src/components/postCards/QuestionCardList.jsx
@@ -1,13 +1,14 @@
 import { Text, TextType } from 'components/text/Text.jsx';
 import * as S from 'components/postCards/questionCards.style.jsx';
 import QuestionCardItem from 'components/postCards/QuestionCardItem';
+import FloatingButton from 'components/floatingButton/FloatingButton.jsx';
 import { postCreateReaction } from 'pages/post/postPage.js';
 import { useState } from 'react';
 import useAsync from 'hooks/useAsync.js';
 
 const REACTION_MAX_INT = 2147483647;
 
-const QuestionCardList = ({ questionInfo, subjectOwner }) => {
+const QuestionCardList = ({ questionInfo, subjectOwner, openModal }) => {
   const { results, count } = questionInfo;
   const [questions, setQuestions] = useState(results);
 
@@ -35,26 +36,31 @@ const QuestionCardList = ({ questionInfo, subjectOwner }) => {
     localStorage.setItem(type, JSON.stringify(localStorageReaction));
   };
   return (
-    <S.PostCardList>
-      <S.PostCardListTitleBox>
-        <S.SpeechBubble />
-        <Text
-          $normalType={TextType.Body1Bol}
-          $mobileType={TextType.Body2Bol}
-          text={`${count}개의 질문이 있습니다.`}
-        />
-      </S.PostCardListTitleBox>
-      {questions?.map((question, questionIndex) => (
-        <QuestionCardItem
-          key={question.id}
-          question={question}
-          setQuestions={setQuestions}
-          subjectOwner={subjectOwner}
-          questionIndex={questionIndex}
-          handleReaction={handleReaction}
-        />
-      ))}
-    </S.PostCardList>
+    <>
+      <S.FloatingButtonItem>
+        <FloatingButton type="W" onClick={openModal} />
+      </S.FloatingButtonItem>
+      <S.PostCardList>
+        <S.PostCardListTitleBox>
+          <S.SpeechBubble />
+          <Text
+            $normalType={TextType.Body1Bol}
+            $mobileType={TextType.Body2Bol}
+            text={`${count}개의 질문이 있습니다.`}
+          />
+        </S.PostCardListTitleBox>
+        {questions?.map((question, questionIndex) => (
+          <QuestionCardItem
+            key={question.id}
+            question={question}
+            setQuestions={setQuestions}
+            subjectOwner={subjectOwner}
+            questionIndex={questionIndex}
+            handleReaction={handleReaction}
+          />
+        ))}
+      </S.PostCardList>
+    </>
   );
 };
